% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{French names}
\usage{
draw_a_name(the_name, the_sex)
}
\arguments{
\item{the_name}{first name of French people}

\item{the_sex}{gender}
}
\description{
French names
}
