% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Saving a dataframe as csv}
\usage{
save_as_csv(df, file_name)
}
\arguments{
\item{df}{dataframe}

\item{file_name}{name of the file}
}
\description{
Saving a dataframe as csv
}
